agent-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weave==0.50.0
neo4j==5.14.0
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2

